requirement.txt
pytest
requests
pyyaml
faker
requests-oauthlib